Add new animations
