+++ src/main/java/cc/polyfrost/oneconfig/gui/animations/circle/EaseInCirc.java
@@ -1,0 +1,50 @@
+/*
+ * This file is part of OneConfig.
+ * OneConfig - Next Generation Config Library for Minecraft: Java Edition
+ * Copyright (C) 2021, 2022 Polyfrost.
+ *   <https://polyfrost.cc> <https://github.com/Polyfrost/>
+ *
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ *   OneConfig is licensed under the terms of version 3 of the GNU Lesser
+ * General Public License as published by the Free Software Foundation, AND
+ * under the Additional Terms Applicable to OneConfig, as published by Polyfrost,
+ * either version 1.0 of the Additional Terms, or (at your option) any later
+ * version.
+ *
+ *   This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ *   You should have received a copy of the GNU Lesser General Public
+ * License.  If not, see <https://www.gnu.org/licenses/>. You should
+ * have also received a copy of the Additional Terms Applicable
+ * to OneConfig, as published by Polyfrost. If not, see
+ * <https://polyfrost.cc/legal/oneconfig/additional-terms>
+ */
+
+package cc.polyfrost.oneconfig.gui.animations.circle;
+
+import cc.polyfrost.oneconfig.gui.animations.Animation;
+
+public class EaseInCirc extends Animation {
+    private static final double CONSTANT_1 = 1.7;
+    private static final double CONSTANT_2 = 2.7;
+
+    /**
+     * @param duration The duration of the animation
+     * @param start    The start of the animation
+     * @param end      The end of the animation
+     * @param reverse  Reverse the animation
+     */
+    public EaseInCirc(int duration, float start, float end, boolean reverse) {
+        super(duration, start, end, reverse, true);
+    }
+
+    // Courtesy of https://easings.net/
+    @Override
+    protected float animate(float x) {
+        return (float) (1 - Math.sqrt(1 - Math.pow(x, 2)));
+    }
+}
